+++ src/docbookgen.h
@@ -320,7 +320,7 @@ class DocbookGenerator final : public OutputGenerator, public OutputGenIntf
     void startPlainFile(const DString &name) override { OutputGenerator::startPlainFile(name); }
     void endPlainFile() override { OutputGenerator::endPlainFile(); }
 
-    void startEmbeddedDoc(int) override {}
+    void startEmbeddedDoc(size_t) override {}
     void endEmbeddedDoc() override {}
     static DString convertToDocbook(const DString &s, bool retainNewline = false, bool citeEntry = false);
